refactor the old impl
